hirc-duckdb-demo/pyproject.toml
[project]
name = "hirc-duckdb-demo"
version = "0.1.0"
description = "Query Snowflake Iceberg tables with DuckDB via Horizon Catalog"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "duckdb>=1.0.0",
    "python-dotenv>=1.0.0",
    "snowflake-cli-labs>=3.0.0",
]

[project.scripts]
hirc-demo-setup = "hirc_demo.cli:setup"
hirc-demo-data = "hirc_demo.cli:load_data"
hirc-demo-rbac = "hirc_demo.cli:grant_rbac"
hirc-demo-cleanup = "hirc_demo.cli:cleanup"

[project.optional-dependencies]
notebook = [
    "jupyter>=1.0.0",
    "pandas>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hirc_demo"]
